requirements.txt
fastapi==0.115.5
pydantic==2.9.2
pydantic-settings==2.6.1
uvicorn==0.32.0
SQLAlchemy==2.0.36
alembic==1.13.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0


# alembic==1.13.3
# annotated-types==0.7.0
# anyio==4.11.0
# bcrypt==4.0.1
# click==8.3.0
# colorama==0.4.6
# dnspython==2.8.0
# email_validator==2.1.1
# fastapi==0.115.5
# greenlet==3.2.4
# h11==0.16.0
# httptools==0.7.1
# idna==3.11
# Mako==1.3.10
# MarkupSafe==3.0.3
# passlib==1.7.4
# pip==25.3
# pydantic==2.9.2
# pydantic_core==2.23.4
# pydantic-settings==2.6.1
# PyJWT==2.9.0
# python-dotenv==1.0.1
# PyYAML==6.0.3
# setuptools==80.9.0
# sniffio==1.3.1
# SQLAlchemy==2.0.36
# starlette==0.41.3
# typing_extensions==4.15.0
# uvicorn==0.32.0
# watchfiles==1.1.1
# websockets==15.0.1
# wheel==0.45.1
